feat: checkout from list
