+++ client-app/pages/account/list-details.vue
@@ -18,11 +18,22 @@
           <VcButton
             :disabled="loading || !pagedListItems.length"
             size="sm"
+            variant="outline"
             prepend-icon="cart"
             class="w-full md:order-last md:w-auto"
             @click="addAllListItemsToCart"
           >
             {{ $t("shared.wishlists.list_details.add_all_to_cart_button") }}
+          </VcButton>
+
+          <VcButton
+            :disabled="loading || !pagedListItems.length"
+            size="sm"
+            prepend-icon="cart-check"
+            class="w-full md:order-last md:w-auto"
+            @click="buyNow"
+          >
+            {{ $t("common.buttons.go_to_checkout") }}
           </VcButton>
 
           <VcButton
@@ -115,7 +126,7 @@ import { breakpointsTailwind, useBreakpoints } from "@vueuse/core";
 import { cloneDeep, isEqual, keyBy, pick } from "lodash";
 import { computed, ref, watchEffect, defineAsyncComponent } from "vue";
 import { useI18n } from "vue-i18n";
-import { onBeforeRouteLeave, onBeforeRouteUpdate } from "vue-router";
+import { onBeforeRouteLeave, onBeforeRouteUpdate, useRouter } from "vue-router";
 import { useAnalytics, useHistoricalEvents, usePageHead } from "@/core/composables";
 import { useAnalyticsUtils } from "@/core/composables/useAnalyticsUtils";
 import { useModuleSettings } from "@/core/composables/useModuleSettings";
@@ -173,6 +184,8 @@ const { pushHistoricalEvent } = useHistoricalEvents();
 usePageHead({
   title: computed(() => t("pages.account.list_details.meta.title", [list.value?.name])),
 });
+
+const router = useRouter();
 
 const { continue_shopping_link } = getModuleSettings({
   [MODULE_XAPI_KEYS.CONTINUE_SHOPPING_LINK]: "continue_shopping_link",
@@ -357,6 +370,10 @@ function selectItemEvent(item: Product | undefined): void {
   }
 
   analytics("selectItem", item, wishlistListProperties.value);
+}
+
+function buyNow() {
+  void router.push({ name: "CartShared", params: { cartId: list.value?.id } });
 }
 
 onBeforeRouteLeave(canChangeRoute);
