+++ locales/en.json
@@ -78,7 +78,8 @@
       "scroll_to_top": "Scroll to top",
       "show_less": "Show less",
       "expand_all": "Expand all",
-      "all_products": "All products"
+      "all_products": "All products",
+      "buy_now": "Buy now"
     },
     "labels": {
       "actions": "Actions",
@@ -273,7 +274,8 @@
       "change_role": "Change role",
       "save_changes": "Save changes",
       "password_expired": "Password is about to expire",
-      "error_details": "Error details"
+      "error_details": "Error details",
+      "buy_now": "Buy now"
     },
     "roles": {
       "org-maintainer": "Organization maintainer",
